Include `node.text` when helpful
Without specifying `node.text`, error messages might not always contain
helpful information.

For example, the following was raised for an `<a>` tag without an
`href`:

```
A11yError:  has a non-meaningful href, it should point to an actual
link.
```

When `${node}` replaced with `${node.text}`, the following was produced:

```
A11yError: "Waiting for a submission..." has a non-meaningful href, it
should point to an actual link.
```