update README to combine axeOptions examples
I changed the rule because it didn't make sense to disable color-contrast, and to provide an option for its check.  Also, I changed it to enabling a rule that is disabled by default, because I think it's better to show an example of enabling additional a11y checks rather than disabling.