Fix cannot find module `qunit` under embroider
When parent applications are using ember-qunit > 5 this causes an
invalid lookup for qunit in requirejs because ember-qunit no longer
provides the AMD shim for qunit and instead uses ember-auto-import. This
change conditionally uses ember-auto-import if that shim is present or
not along with adding test cases for general embroider compatibility.