use Testem.afterTests instead of QUnit.done
Sometimes, like in CI environments, there can exist a race condition
where Testem closes before the local middleware reporter can finish
writing the reports to the filesystem.

According to [Testem's README][1], it appears the proper way to do
this is to use `Testem.afterTests` to send reports. This will keep testem
open and prevent the process from exiting before the reports are written.

This seems to be a [known issue with Testem itself][2].

[1]: https://github.com/testem/testem/blob/0bd05cec7f83bfe9f2e10ddfab54556d724ee179/README.md#running-browser-code-after-tests-complete
[2]: https://github.com/testem/testem/issues/1577